+++ typed_sql/test/typed_sql/foreign_key/nullable_composite_foreign_key/nullable_composite_foreign_key_test.dart
@@ -142,9 +142,7 @@ void main() {
 
   r.addTest('books.where(.author.firstName.equals(Bucks))', (db) async {
     final result = await db.books
-        .where(
-          (b) => b.author.firstName.equalsUnlessNull(toExpr('Bucks')),
-        )
+        .where((b) => b.author.firstName.equalsValue('Bucks'))
         .fetch();
     check(result).length.equals(2);
   });
@@ -246,8 +244,8 @@ void main() {
               db.books
                   .where(
                     (b) =>
-                        b.authorFirstName.equalsUnlessNull(author.firstName) &
-                        b.authorLastName.equalsUnlessNull(author.lastName),
+                        b.authorFirstName.equals(author.firstName) &
+                        b.authorLastName.equals(author.lastName),
                   )
                   .select((b) => (b.stock,))
                   .sum()
@@ -270,8 +268,8 @@ void main() {
         .join(db.books)
         .on(
           (author, book) =>
-              book.authorFirstName.equalsUnlessNull(author.firstName) &
-              book.authorLastName.equalsUnlessNull(author.lastName),
+              book.authorFirstName.equals(author.firstName) &
+              book.authorLastName.equals(author.lastName),
         )
         .groupBy((author, book) => (author,))
         .aggregate(